+++ packages/backend/src/routers/apiV1Router.ts
@@ -1,3 +1,8 @@
+import {
+    getPasswordSchema,
+    ParameterError,
+    validatePassword,
+} from '@lightdash/common';
 import express from 'express';
 import passport from 'passport';
 import { lightdashConfig } from '../config/lightdashConfig';
@@ -52,6 +57,10 @@ apiV1Router.get('/flash', (req, res) => {
 
 apiV1Router.post('/register', unauthorisedInDemo, async (req, res, next) => {
     try {
+        if (req.body.password && !validatePassword(req.body.password)) {
+            next(new ParameterError('Password does not meet requirements'));
+            return;
+        }
         const lightdashUser = await userService.registerUser({
             firstName: sanitizeStringParam(req.body.firstName),
             lastName: sanitizeStringParam(req.body.lastName),
@@ -146,6 +155,21 @@ apiV1Router.get(
         scope: ['profile', 'email'],
     }),
 );
+apiV1Router.get(
+    '/login/gdrive',
+    storeOIDCRedirect,
+    passport.authenticate('google', {
+        scope: [
+            'profile',
+            'email',
+            'https://www.googleapis.com/auth/drive.metadata.readonly',
+        ],
+        accessType: 'offline',
+        prompt: 'consent',
+        session: false,
+        includeGrantedScopes: true,
+    }),
+);
 
 apiV1Router.get(
     lightdashConfig.auth.google.callbackPath,
@@ -153,6 +177,7 @@ apiV1Router.get(
         failureRedirect: '/api/v1/oauth/failure',
         successRedirect: '/api/v1/oauth/success',
         failureFlash: true,
+        includeGrantedScopes: true,
     }),
 );
 apiV1Router.get('/oauth/failure', redirectOIDCFailure);