+++ packages/frontend/src/components/DashboardFilter/FilterConfiguration/FilterConfiguration.styled.tsx
@@ -1,8 +1,0 @@
-import styled from 'styled-components';
-
-export const ConfigureFilterWrapper = styled.div`
-    height: 100%;
-    display: flex;
-    flex-direction: column;
-    gap: 15px;
-`;